+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Change the width of the catalog table columns so they evenly spread.
 - Fix catalog table 'Latest release' column sorting.
 - Fix catalog table 'Last released' column sorting.
+- Disable 'unsorted' click for catalog table order buttons.
 
 ## [0.20.0] - 2024-04-04