+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.20.1] - 2024-04-08
+
 ### Fixed
 
 - Change the width of the catalog table columns so they evenly spread.
@@ -443,7 +445,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Disable anonymous access.
 
-[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.20.0...HEAD
+[Unreleased]: https://github.com/giantswarm/backstage/compare/v0.20.1...HEAD
+[0.20.1]: https://github.com/giantswarm/backstage/compare/v0.20.0...v0.20.1
 [0.20.0]: https://github.com/giantswarm/backstage/compare/v0.19.0...v0.20.0
 [0.19.0]: https://github.com/giantswarm/backstage/compare/v0.18.1...v0.19.0
 [0.18.1]: https://github.com/giantswarm/backstage/compare/v0.18.0...v0.18.1